requirements.txt
gunicorn
Django
django-storages[azure]
azure-storage-common
sorl-thumbnail
django-filer
psycopg2-binary==2.8.6
djangorestframework
django-oauth-toolkit
django-cors-middleware


